Feat: condition to stop the game added
My last commit about this project what about a condition to block the
loop iteration through buttons. But I had to create an other repository
on gitHub because I was not able to fix it. My css was not added and due
to this, the website was not displayed as I wanted to.